% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{athena_temp_db}
\alias{athena_temp_db}
\title{athena_temp_db}
\usage{
athena_temp_db(conn, check_exists = TRUE)
}
\arguments{
\item{conn}{This is a connection object returned by \code{connect_athena()}.}

\item{check_exists}{This is \code{TRUE} by default.  If set to \code{FALSE} then it will not check if the temporary database exists and will not create it if it does not exist.}
}
\description{
Returns a string containing the temporary database name for an Athena Connection.
Will also check if that temporary database exists and if not will create it, unless \code{check_exists = FALSE}.
}
\examples{
 con <- connect_athena() # creates a connection with sensible defaults
 athena_temp_db(con)
 > [1] "mojap_de_temp_alpha_user_Rdbtools_User"

}
