% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{create_temp_table}
\alias{create_temp_table}
\title{create_temp_table (deprecated)}
\usage{
create_temp_table(sql, table_name, conn = NULL)
}
\arguments{
\item{sql}{A sql command which generates a table to be created in the temporary location.}

\item{table_name}{The name of the table to be created in the user's temporary database.}

\item{conn}{(optional) A DBIConnection object, as returned by \code{connect_athena()}. If unused then the query will create its own connection and close it subsequently.  Note that if you choose not to supply this argument then you may find you cannot access the table later with a different connection.}
}
\description{
A convenience function to match dbtools which creates a table in the user's temporary database.
Please now use \code{\link[DBI:dbExecute]{dbExecute()}} directly instead, e.g. with variables \code{conn}, \code{table_name} and \code{sql} then run: \verb{dbExecute(conn, glue("CREATE TABLE __temp__.\{table_name\} as \{sql\}")})
}
