% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbExecute,MoJAthenaConnection,character-method}
\alias{dbExecute,MoJAthenaConnection,character-method}
\title{dbExecute}
\usage{
\S4method{dbExecute}{MoJAthenaConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}
}
\description{
See \link[noctua:Query]{noctua::Query}. This function calls \code{noctua::dbExecute()}, after replacing any references to \verb{__temp__}
in the statement with your temporary database in Athena. Your temporary database will be created
if you do not already have one.
}
