% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbGetTables,MoJAthenaConnection-method}
\alias{dbGetTables,MoJAthenaConnection-method}
\title{dbGetTables}
\usage{
\S4method{dbGetTables}{MoJAthenaConnection}(conn, schema = NULL, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{schema}{Athena schema, default set to NULL to return all tables from all Athena schemas.
Note: The use of DATABASE and SCHEMA is interchangeable within Athena.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \link[noctua:AthenaConnection]{noctua::dbGetTables}. This function calls \code{noctua::dbGetTables()} but if
the schema argument is \verb{__temp__} then it looks at your temporary database in Athena.
NOTE: that the schema is the first argument here, to retain backwards compatibility with previous version of \code{Rdbtools}.
From \code{noctua} 2.6.2 the first argument is \code{catalog}.
}
