% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_athena_credentials}
\alias{get_athena_credentials}
\title{get_athena_credentials}
\usage{
get_athena_credentials(
  aws_region = NULL,
  session_duration = 3600L,
  role_session_name = NULL
)
}
\description{
Gets temporary credentials for an AWS service
}
