% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_aws_credentials}
\alias{get_aws_credentials}
\title{get_aws_credentials}
\usage{
get_aws_credentials(
  aws_region = NULL,
  session_duration = 3600L,
  role_session_name = NULL,
  ...
)
}
\arguments{
\item{aws_region}{Default is NULL, which will look for relevant environment variables and if not found set this to be the relevant region for most Analytical Platform users.}

\item{session_duration}{The number of seconds which the session should last before needing new authentication. Minimum of 900.}

\item{role_session_name}{This is a parameter for authentication, and should be left to NULL in normal operation.}

\item{...}{Further arguments for the \code{paws} function \code{assume_role_with_web_identity}}
}
\description{
Gets temporary credentials for an AWS service.
}
