% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dbDisconnect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{noctua}{\code{\link[noctua:dbDisconnect]{dbDisconnect}}}
}}

