% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_connection.R
\name{refresh_athena_connection}
\alias{refresh_athena_connection}
\title{refresh_athena_connection}
\usage{
refresh_athena_connection(conn)
}
\arguments{
\item{conn}{This is the connection which will be refreshed.}
}
\description{
Refreshes an athena connection to the AP (e.g. if the credentials have expired).
}
\examples{
 con <- connect_athena() # creates a connection with sensible defaults
 data <- dbGetQuery(con, "SELECT * FROM database.table") # queries and puts data in R environment
 # Some time later...
 refresh_athena_connection(con) # refresh the connection for any further queries on the same session
 dbDisconnect(con) # disconnects the connection
}
