% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_connection.R
\name{refresh_if_expired}
\alias{refresh_if_expired}
\title{refresh_if_expired}
\usage{
refresh_if_expired(conn, window = 5 * 60)
}
\arguments{
\item{conn}{This is the connection which has expired, but will be refreshed.}

\item{window}{The number of seconds in advance of expiry that a refresh will still happen (default 5 mins).}
}
\description{
Refreshes an athena connection to the AP only if the credentials have expired.
}
