% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{sqlCreateTable,MoJAthenaConnection-method}
\alias{sqlCreateTable,MoJAthenaConnection-method}
\title{#' dbWriteTable
#'
#' @rdname dbWriteTable
#' @export
setMethod("dbWriteTable", c("MoJAthenaConnection", "character", "data.frame"),
function(conn, name, value, overwrite=FALSE, append=FALSE,
row.names = NA, field.types = NULL,
partition = NULL, s3.location = NULL, file.type = c("tsv", "csv", "parquet", "json"),
compress = FALSE, max.batch = Inf, ...) {
# prepare the statement
name <- stringr::str_replace_all(name, "\strong{temp}", conn@MoJdetails$temp_db_name)
# run the query (as AthenaConnection to avoid recursion)
resp <- dbWriteTable(as(conn, "AthenaConnection"),
name, value, overwrite, append,
row.names, field.types,
partition, s3.location, file.type,
compress, max.batch, ...)
return(resp)
}
)}
\usage{
\S4method{sqlCreateTable}{MoJAthenaConnection}(
  con,
  table,
  fields,
  field.types = NULL,
  partition = NULL,
  s3.location = NULL,
  file.type = c("tsv", "csv", "parquet", "json"),
  compress = FALSE,
  ...
)
}
\description{
#' dbWriteTable
#'
#' @rdname dbWriteTable
#' @export
setMethod("dbWriteTable", c("MoJAthenaConnection", "character", "data.frame"),
function(conn, name, value, overwrite=FALSE, append=FALSE,
row.names = NA, field.types = NULL,
partition = NULL, s3.location = NULL, file.type = c("tsv", "csv", "parquet", "json"),
compress = FALSE, max.batch = Inf, ...) {
# prepare the statement
name <- stringr::str_replace_all(name, "\strong{temp}", conn@MoJdetails$temp_db_name)
# run the query (as AthenaConnection to avoid recursion)
resp <- dbWriteTable(as(conn, "AthenaConnection"),
name, value, overwrite, append,
row.names, field.types,
partition, s3.location, file.type,
compress, max.batch, ...)
return(resp)
}
)
}
