% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{write_small_temp_table}
\alias{write_small_temp_table}
\title{write_small_temp_table}
\usage{
write_small_temp_table(
  con,
  table_name,
  table_to_write,
  overwrite = FALSE,
  append = FALSE,
  row.names = NA
)
}
\description{
I can't get dbWriteTable to work, so this is a hacky way to create a table
and run INSERT INTO to append values to it.
Probably only works with numbers and strings
Probably won't work with more than a few hundred values.
Will only write to tempdb, but don't put it in the table_name parameter.
}
