% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbExecute,MoJAthenaConnection,character-method}
\alias{dbExecute,MoJAthenaConnection,character-method}
\title{dbExecute}
\usage{
\S4method{dbExecute}{MoJAthenaConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{statement}{a character string containing SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \link[noctua:Query]{noctua::Query}
}
