% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbExecute,MoJAthenaConnection,character-method}
\alias{dbExecute,MoJAthenaConnection,character-method}
\title{dbExecute}
\usage{
\S4method{dbExecute}{MoJAthenaConnection,character}(conn, statement, ...)
}
\description{
dbExecute
}
