% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbExistsTable,MoJAthenaConnection,character-method}
\alias{dbExistsTable,MoJAthenaConnection,character-method}
\title{dbExistsTable}
\usage{
\S4method{dbExistsTable}{MoJAthenaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}
}
\description{
See \code{\link[noctua:dbExistsTable]{noctua::dbExistsTable()}}. This function calls noctua::dbExistsTable(), after replacing any references to \verb{__temp__}
in the name argument with your temporary database in Athena.
}
