% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbExistsTable,MoJAthenaConnection,character-method}
\alias{dbExistsTable,MoJAthenaConnection,character-method}
\title{dbExistsTable}
\usage{
\S4method{dbExistsTable}{MoJAthenaConnection,character}(conn, name, ...)
}
\description{
dbExistsTable
}
