% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbGetQuery,MoJAthenaConnection,character-method}
\alias{dbGetQuery,MoJAthenaConnection,character-method}
\title{dbGetQuery}
\usage{
\S4method{dbGetQuery}{MoJAthenaConnection,character}(conn, statement, statistics = FALSE, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}
}
\description{
See \link[noctua:AthenaConnection]{noctua::dbGetQuery}. This function calls \code{noctua::dbGetQuery()}, after replacing any references to \verb{__temp__}
in the statement with your temporary database in Athena. Your temporary database will be created
if you do not already have one.
}
