% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbGetQuery,MoJAthenaConnection,character-method}
\alias{dbGetQuery,MoJAthenaConnection,character-method}
\title{dbGetQuery}
\usage{
\S4method{dbGetQuery}{MoJAthenaConnection,character}(conn, statement, statistics = FALSE, ...)
}
\description{
dbGetQuery
}
