% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbGetQuery,MoJAthenaConnection,character-method}
\alias{dbGetQuery,MoJAthenaConnection,character-method}
\title{dbGetQuery}
\usage{
\S4method{dbGetQuery}{MoJAthenaConnection,character}(conn, statement, statistics = FALSE, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{statement}{a character string containing SQL.}

\item{statistics}{If set to \code{TRUE} will print out AWS Athena statistics of query.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \code{\link[noctua:dbGetQuery]{noctua::dbGetQuery()}}
}
