% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbGetTables,MoJAthenaConnection-method}
\alias{dbGetTables,MoJAthenaConnection-method}
\title{dbGetTables}
\usage{
\S4method{dbGetTables}{MoJAthenaConnection}(conn, schema = NULL, ...)
}
\description{
dbGetTables
}
