% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbListFields,MoJAthenaConnection,character-method}
\alias{dbListFields,MoJAthenaConnection,character-method}
\title{dbListFields}
\usage{
\S4method{dbListFields}{MoJAthenaConnection,character}(conn, name, ...)
}
\description{
dbListFields
}
