% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbListFields,MoJAthenaConnection,character-method}
\alias{dbListFields,MoJAthenaConnection,character-method}
\title{dbListFields}
\usage{
\S4method{dbListFields}{MoJAthenaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{name}{a character string with the name of the remote table.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \code{\link[noctua:dbListFields]{noctua::dbListFields()}}. This function calls \code{noctua::dbListFields()}, after replacing any references to \verb{__temp__}
in the name argument with your temporary database in Athena.
}
