% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbListFields,MoJAthenaConnection,character-method}
\alias{dbListFields,MoJAthenaConnection,character-method}
\title{dbListFields}
\usage{
\S4method{dbListFields}{MoJAthenaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \code{\link[noctua:dbListFields]{noctua::dbListFields()}}. This function calls \code{noctua::dbListFields()}, after replacing any references to \verb{__temp__}
in the name argument with your temporary database in Athena.
}
