% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbListTables,MoJAthenaConnection-method}
\alias{dbListTables,MoJAthenaConnection-method}
\title{dbListTables}
\usage{
\S4method{dbListTables}{MoJAthenaConnection}(conn, schema = NULL, ...)
}
\description{
dbListTables
}
