% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbListTables,MoJAthenaConnection-method}
\alias{dbListTables,MoJAthenaConnection-method}
\title{dbListTables}
\usage{
\S4method{dbListTables}{MoJAthenaConnection}(conn, schema = NULL, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{schema}{Athena schema, default set to NULL to return all tables from all Athena schemas.
Note: The use of DATABASE and SCHEMA is interchangeable within Athena.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \code{\link[noctua:dbListTables]{noctua::dbListTables()}}. This function calls \code{noctua::dbListTables()} but if
the schema argument is \verb{__temp__} then it looks at your temporary database in Athena.
}
