% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbRemoveTable,MoJAthenaConnection,character-method}
\alias{dbRemoveTable,MoJAthenaConnection,character-method}
\title{dbRemoveTable}
\usage{
\S4method{dbRemoveTable}{MoJAthenaConnection,character}(conn, name, delete_data = TRUE, confirm = FALSE, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}
}
\description{
See \code{\link[noctua:dbRemoveTable]{noctua::dbRemoveTable()}}. This function calls \code{noctua::dbRemoveTable()}, after replacing any references to \verb{__temp__}
in the names argument with your temporary database in Athena.
}
