% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbRemoveTable,MoJAthenaConnection,character-method}
\alias{dbRemoveTable,MoJAthenaConnection,character-method}
\title{dbRemoveTable}
\usage{
\S4method{dbRemoveTable}{MoJAthenaConnection,character}(conn, name, delete_data = TRUE, confirm = FALSE, ...)
}
\description{
dbRemoveTable
}
