% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbRemoveTable,MoJAthenaConnection,character-method}
\alias{dbRemoveTable,MoJAthenaConnection,character-method}
\title{dbRemoveTable}
\usage{
\S4method{dbRemoveTable}{MoJAthenaConnection,character}(conn, name, delete_data = TRUE, confirm = FALSE, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{connect_athena()}}

\item{name}{A character string specifying a DBMS table name.}

\item{delete_data}{Deletes S3 files linking to AWS Athena table}

\item{confirm}{Allows for S3 files to be deleted without the prompt check. It is recommend to leave this set to \code{FALSE}
to avoid deleting other S3 files when the table's definition points to the root of S3 bucket.}

\item{...}{Other parameters passed on to methods.}
}
\description{
See \code{\link[noctua:dbRemoveTable]{noctua::dbRemoveTable()}}. This function calls \code{noctua::dbRemoveTable()}, after replacing any references to \verb{__temp__}
in the names argument with your temporary database in Athena.
}
