% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{dbWriteTable,MoJAthenaConnection,character,data.frame-method}
\alias{dbWriteTable,MoJAthenaConnection,character,data.frame-method}
\title{dbWriteTable}
\usage{
\S4method{dbWriteTable}{MoJAthenaConnection,character,data.frame}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  row.names = NA,
  field.types = NULL,
  partition = NULL,
  s3.location = NULL,
  file.type = c("tsv", "csv", "parquet", "json"),
  compress = FALSE,
  max.batch = Inf,
  ...
)
}
\description{
dbWriteTable
}
