% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdbtools.R
\docType{package}
\name{Rdbtools-package}
\alias{Rdbtools}
\alias{Rdbtools-package}
\title{Rdbtools: Connects the MoJ Analytical Platform to Athena}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
