% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdbtools-package.R, R/Rdbtools.R
\docType{package}
\name{Rdbtools-package}
\alias{Rdbtools}
\alias{Rdbtools-package}
\title{Rdbtools: Connects the MoJ Analytical Platform to Athena}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
