% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_commands.R
\name{read_sql}
\alias{read_sql}
\title{read_sql (deprecated)}
\usage{
read_sql(sql_query, return_df_as = "tibble", conn = NULL)
}
\arguments{
\item{sql_query}{A sql command which is passed to \code{\link[=dbGetQuery]{dbGetQuery()}}}

\item{return_df_as}{must be 'tibble' or 'data.table'.}

\item{conn}{(optional) A DBIConnection object, as returned by \code{connect_athena()}. If unused then the query will create its own connection and close it subsequently. If reading a table created by a previous connection then the same connection must be supplied, otherwise you may get permission errors.}
}
\description{
A convenience function to match dbtools which reads a table into a tibble or dataframe.
Please now use \code{\link[=dbGetQuery]{dbGetQuery()}} directly instead.
}
